Patch: move MeshBase to libnomadEval
MeshBase is used in EvalPoint (libnomadEval). Some compilers are more picky.
