+++ CMakeLists.txt
@@ -7,7 +7,7 @@ cmake_minimum_required(VERSION 3.14 FATAL_ERROR)
 
 set(NOMAD_VERSION_MAJOR 4)
 set(NOMAD_VERSION_MINOR 3)
-set(NOMAD_VERSION_PATCH 0)  
+set(NOMAD_VERSION_PATCH 1)  
 set(NOMAD_VERSION ${NOMAD_VERSION_MAJOR}.${NOMAD_VERSION_MINOR}.${NOMAD_VERSION_PATCH})
 
 # name of the project
