+++ src/Algos/DMultiMads/DMultiMadsIteration.hpp
@@ -48,11 +48,10 @@
 #define __NOMAD_4_3_DMULTIMADSITERATION__
 
 #include "../../Algos/Iteration.hpp"
-#include "../../Algos/MeshBase.hpp"
 #include "../../Algos/DMultiMads/DMultiMadsUpdate.hpp"
 #include "../../Algos/Mads/Poll.hpp"
 #include "../../Algos/Mads/Search.hpp"
-
+#include "../../Eval/MeshBase.hpp"
 
 #include "../../nomad_nsbegin.hpp"
