+++ src/nomad_version.hpp
@@ -51,4 +51,4 @@
 
 #undef NOMAD_VERSION_NUMBER
 /// Define version number
-#define NOMAD_VERSION_NUMBER "4.3.0"
+#define NOMAD_VERSION_NUMBER "4.3.1"
